requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
black==19.3b0
Brlapi==0.6.6
camelot-py==0.7.3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
command-not-found==0.3
cozmo==1.4.10
cozmoclad==3.4.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
distro==1.4.0
Elixir==0.7.1
et-xmlfile==1.0.1
ghostscript==0.6
httplib2==0.9.2
idna==2.8
importlib-metadata==0.22
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
lazy-object-proxy==1.4.2
louis==3.5.0
macaroonbakery==1.1.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
minecart==0.3.0
more-itertools==7.2.0
netifaces==0.10.4
numpy==1.17.2
olefile==0.45.1
onboard==1.4.1
opencv-python==4.1.1.26
openpyxl==2.6.3
packaging==19.1
pandas==0.25.1
parso==0.5.1
pbr==5.4.3
pdfminer3==2018.12.3.0
pdfminer3k==1.3.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.0
ply==3.11
prompt-toolkit==2.0.9
protobuf==3.0.0
ptyprocess==0.6.0
py==1.8.0
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycryptodome==3.9.0
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
PyPDF2==1.26.0
pyRFC3339==1.0
pytest==5.1.2
python-apt===1.6.4-elementary6-ubuntu5.0.1
python-dateutil==2.8.0
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
system-service==0.3
tabula-py==1.4.1
Tempita==0.5.2
toml==0.10.0
traitlets==4.3.2
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.3
virtualenv==16.7.5
wcwidth==0.1.7
wrapt==1.11.2
xkit==0.0.0
xlrd==0.7.1
xlwt==0.7.2
zipp==0.6.0
